requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-SQLAlchemy

requests
beautifulsoup4
Pillow
